Extract ForgetfulMap to own package
