+++ lib/dedupe/main.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"time"
 
+	"plane.watch/lib/dedupe/forgetfulmap"
 	"plane.watch/lib/tracker"
 	"plane.watch/lib/tracker/beast"
 	"plane.watch/lib/tracker/mode_s"
@@ -19,13 +20,13 @@ Consider a message a duplicate if we have seen it in the last minute
 type (
 	Filter struct {
 		events chan tracker.Event
-		list   *ForgetfulSyncMap
+		list   *forgetfulmap.ForgetfulSyncMap
 	}
 )
 
 func NewFilter() *Filter {
 	return &Filter{
-		list:   NewForgetfulSyncMap(10*time.Second, 60*time.Second),
+		list:   forgetfulmap.NewForgetfulSyncMap(10*time.Second, 60*time.Second),
 		events: make(chan tracker.Event),
 	}
 }
